+++ src/main/java/roomescape/user/controller/ReservationTimeController.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.user.controller;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -10,9 +10,9 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.domain.ReservationTime;
-import roomescape.dto.ReservationTimeRequest;
-import roomescape.dto.ReservationTimeResponse;
-import roomescape.service.ReservationTimeService;
+import roomescape.user.dto.ReservationTimeRequest;
+import roomescape.user.dto.ReservationTimeResponse;
+import roomescape.user.service.ReservationTimeService;
 
 @RestController
 public class ReservationTimeController {
