+++ src/main/java/roomescape/user/dao/ReservationDao.java
@@ -1,4 +1,4 @@
-package roomescape.dao;
+package roomescape.user.dao;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
